+++ lib/src/views/settings/privacy.view.dart
@@ -3,6 +3,7 @@ import 'package:go_router/go_router.dart';
 import 'package:twonly/globals.dart';
 import 'package:twonly/src/constants/routes.keys.dart';
 import 'package:twonly/src/utils/misc.dart';
+import 'package:twonly/src/utils/storage.dart';
 
 class PrivacyView extends StatefulWidget {
   const PrivacyView({super.key});
@@ -15,6 +16,20 @@ class _PrivacyViewState extends State<PrivacyView> {
   @override
   void initState() {
     super.initState();
+  }
+
+  Future<void> toggleAuthRequirementOnStartup() async {
+    final isAuth = await authenticateUser(
+      gUser.screenLockEnabled
+          ? context.lang.settingsScreenLockAuthMessageDisable
+          : context.lang.settingsScreenLockAuthMessageEnable,
+    );
+    if (!isAuth) return;
+    await updateUserdata((u) {
+      u.screenLockEnabled = !u.screenLockEnabled;
+      return u;
+    });
+    setState(() {});
   }
 
   @override
@@ -42,6 +57,15 @@ class _PrivacyViewState extends State<PrivacyView> {
               },
             ),
             onTap: () => context.push(Routes.settingsPrivacyBlockUsers),
+          ),
+          ListTile(
+            title: Text(context.lang.settingsScreenLock),
+            subtitle: Text(context.lang.settingsScreenLockSubtitle),
+            onTap: toggleAuthRequirementOnStartup,
+            trailing: Switch(
+              value: gUser.screenLockEnabled,
+              onChanged: (a) => toggleAuthRequirementOnStartup(),
+            ),
           ),
           ListTile(
             title: Text(context.lang.contactVerifyNumberTitle),